Queue and recycle (reposition) the close-up cubes
Now we'll store our skyline objects in a queue and keep checking
whether the first object in it should be recycled. If so, we'll
reposition it and move it to the back of the queue. Let's use a
recycleOffset variable to configure how far behind Runner this reuse
should occur. A value of 60 seems to work well.