Added a physic material to the runner
We'll use Unity's physics engine to make it jump, fall, and collide
with the platforms, so add a Rigidbody component to Runner via
Component / Physics / Rigidbody. We don't want it to rotate or
disappear into the distance, so constrain it by freezing its Z position
and all rotation axes.
As movement will be accomplished by gliding across the platforms, let's
create a physic material (Create / Physic Material) with no friction
whatsoever. Set all its fields to zero and both combine options to
maximum. This way friction will be determined by whatever it's gliding
across.
Name the new physic material Runner PMat, put it in the Runner folder,
and assign it to the Material field of the Box Collider of Runner.
Reposition Runner to (0, 2, 0) so that it will begin by falling down on
the first platform. Then try out play mode to see what happens!