Only apply acceleration to the runner when in contact with a platform
So Runner falls on the flatworm and then moves to the right. It even
falls again after it moves past the platform. But when it happens to
collide with the side of the next platform it behaves a bit weird. This
is because we're still changing its position in an Update method. We
should leave its movement to the physics engine and instead apply
forces to it.
Remove the call to Translate from the Update method of Runner. Instead,
we'll use two of Unity's collision event methods – OnCollisionEnter and
OnCollisionExit – to detect when we touch or leave a platform. As long
as we're touching a platform, we apply an acceleration to make us run
faster.
Let's make the acceleration configurable and set it to 5 in the editor.