Enabled the runner to jump!
We'll add a variable to Runner so we can configure its jump velocity.
We'll use a vector instead of just a float so we can provide both a
vertical and horizontal component. Set the corresponding field in the
editor to (1, 7, 0).
We want Runner to jump only when it's touching a platform while the
jump button is pressed