Prevent the player saving themselves from a gap
Now we can jump! However, if Runner hits a platform from the side, we
can perform multiple jumps while it's still touching the platform,
launching ourselves out of the gap. To prevent this, we'll decree that
once jumped we are no longer touching the platform, even if we really
are. This allows for one jump after colliding, which usually isn't
enough to escape from a gap.