Hide Runner before game start
We want Runner to be hidden before the first game is started. We can do
this by simply deactivating it in its Start method and activating it
when the game start event is triggered. We'll also remember its
starting position so we can reset it each game start. Let's reset
distanceTraveled too, so it's immediately up to date. (I've omitted the
contents of unmodified methods in the code below.)