Freeze the position of Runner at Game Over
Additionally, let's freeze the position of Runner once it's game over.
We can do this by making its rigid body kinematic, which has the added
benefit of resetting its velocity. Be sure to undo this change when a
new game starts, though. Also, it's undesirable to keep the Update
method running after the game ends, so we should disable the runner
component until a new game begins.