Only display the platforms at game start
Now that Runner behaves, let's modify our platform manager so it reacts
to the events in a similar fashion. It should only be enabled when a
game is in progress and the platforms should only become visible after
the first game start event has been triggered.
We can achieve this by initially instantiating the platforms somewhere
far behind the camera and moving the recyle loop to a new GameStart
method.